# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StudyMind API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects/{subjectId}/uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List uploads for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Upload"
    post:
      operationId: createUpload
      tags: [uploads]
      summary: Create an upload record (text/URL based)
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUploadRequest"
      responses:
        "201":
          description: Created upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
  /subjects/{subjectId}/uploads/{uploadId}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get an upload
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an upload
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /subjects/{subjectId}/uploads/{uploadId}/process:
    post:
      operationId: processUpload
      tags: [uploads]
      summary: Trigger AI processing of an upload
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
        - name: uploadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Processing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingResult"
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams for current user
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{teamId}/members:
    get:
      operationId: listTeamMembers
      tags: [teams]
      summary: List team members
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member to a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTeamMemberRequest"
      responses:
        "201":
          description: Added member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
...
components:
  schemas:
...
    Upload:
      type: object
      properties:
        id:
          type: string
        subjectId:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [pdf, text, url, audio, video, image, word, ppt]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        content:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        processingLog:
          type: string
          nullable: true
        flashcardsGenerated:
          type: integer
        nodesGenerated:
          type: integer
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, subjectId, title, type, status, flashcardsGenerated, nodesGenerated, uploadedBy, createdAt]
    CreateUploadRequest:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [pdf, text, url, audio, video, image, word, ppt]
        content:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
      required: [title, type]
    ProcessingResult:
      type: object
      properties:
        uploadId:
          type: string
        summary:
          type: string
        flashcardsCreated:
          type: integer
        nodesCreated:
          type: integer
        conflictsDetected:
          type: integer
        processingLog:
          type: string
      required: [uploadId, summary, flashcardsCreated, nodesCreated, conflictsDetected, processingLog]
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        ownerId:
          type: string
        memberCount:
          type: integer
        subjectCount:
          type: integer
        inviteCode:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, ownerId, memberCount, subjectCount, inviteCode, createdAt]
    CreateTeamRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [name]
    TeamMember:
      type: object
      properties:
        id:
          type: string
        teamId:
          type: string
        userId:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, editor, viewer]
        joinedAt:
          type: string
          format: date-time
      required: [id, teamId, userId, displayName, role, joinedAt]
    AddTeamMemberRequest:
      type: object
      properties:
        email:
          type: string
        role:
          type: string
          enum: [editor, viewer]
      required: [email, role]
...